infra/requirements-infra.txt
attrs==25.3.0
aws-cdk-lib==2.147.0
aws-cdk.asset-awscli-v1==2.2.246
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.146.0a0
aws-lambda-powertools==2.40.0
boto3==1.34.131
botocore==1.34.162
cattrs==24.1.3
constructs==10.4.2
exceptiongroup==1.3.0
importlib_resources==6.5.2
jmespath==1.0.1
jsii==1.112.0
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.10.4
six==1.17.0
typeguard==2.13.3
typing_extensions==4.14.1
urllib3==2.5.0
